tests/wpt/web-platform-tests/portals/portals-host-null.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  test(t => {
    assert_equals(window.portalHost, null, "window.portalHost should be null");
  });
</script>
</body>
